+++ lib/decentworks/date_support/configuration.rb
@@ -20,10 +20,10 @@ module Decentworks
       end
 
       # 開始月名
-      delegate :beginning_of_first_quarter, to: :configuration
+      delegate :first_quarter_month_name, to: :configuration
 
       # 開始月数
-      delegate :beginning_of_first_quarter_month, to: :configuration
+      delegate :first_quarter_month, to: :configuration
 
       # 設定を初期化する
       def reset_configuration!
@@ -45,28 +45,28 @@ module Decentworks
       # 開始月の初期値（1月）
       DEFAULT = :january
 
-      # 開始月
-      attr_reader :beginning_of_first_quarter
+      # 開始月名
+      attr_reader :first_quarter_month_name
 
       def initialize
-        @beginning_of_first_quarter = DEFAULT
+        @first_quarter_month_name = DEFAULT
       end
 
-      def beginning_of_first_quarter=(month_name)
+      def first_quarter_month_name=(month_name)
         unless MONTHS.key?(month_name)
           raise ::ArgumentError,
-                "beginning_of_first_quarter must be one of #{MONTHS.keys.inspect} (got #{month_name.inspect})"
+                "first_quarter_month_name must be one of #{MONTHS.keys.inspect} (got #{month_name.inspect})"
         end
 
-        @beginning_of_first_quarter = month_name
+        @first_quarter_month_name = month_name
       end
 
       # 開始月数
-      def beginning_of_first_quarter_month = MONTHS.fetch(beginning_of_first_quarter)
+      def first_quarter_month = MONTHS.fetch(first_quarter_month_name)
 
       # 1月始まりとの開始月のずれ
       def first_quarter_month_offset
-        beginning_of_first_quarter_month - 1
+        first_quarter_month - 1
       end
     end
   end